Fixed grunt task
