+++ .jshintrc
@@ -22,6 +22,7 @@
 	"maxcomplexity": 7,
 	"globals": {
 		"angular": false,
+		"Spinner": false,
 		"define": false,
 
 		"jasmine": false,
