+++ package.json
@@ -6,13 +6,19 @@
 		"url": "http://github.com/urish/angular-spinner.git"
 	},
 	"main": "./index.js",
+	"author": "Uri Shaked",
 	"dependencies": {
 		"spin.js": "^2.0.0"
 	},
+	"files": [
+		"src",
+		"dist"
+	],
 	"license": "MIT",
 	"devDependencies": {
+		"coveralls": "2.11.4",
 		"grunt": "0.4.5",
-		"load-grunt-tasks": "3.2.0",
+		"grunt-contrib-concat": "^0.5.1",
 		"grunt-contrib-jshint": "0.11.3",
 		"grunt-contrib-uglify": "0.9.2",
 		"grunt-karma": "0.12.0",
@@ -20,7 +26,7 @@
 		"karma-coverage": "0.5.2",
 		"karma-jasmine": "0.3.6",
 		"karma-phantomjs-launcher": "0.2.1",
-		"coveralls": "2.11.4"
+		"load-grunt-tasks": "3.2.0"
 	},
 	"engines": {
 		"node": ">=0.10.0"
